Implemented DFT convolution algorithm

+++ CPP/include/interpolation.h
@@ -15,9 +15,9 @@
 static const std::vector<int> possible_phi = {-45, -30, -15, 0, 15, 30, 45, 60, 75};
 
 // Functions
-WAVFile hrir_interpolation(int theta, int phi);
+WAVFile hrir_interpolation(double theta, double phi);
 
-WAVFile load_hrir(int theta, int phi);
+WAVFile load_hrir(double theta, double phi);
 
 std::vector<int> get_possible_theta(int phi);
 
@@ -27,16 +27,16 @@ std::vector<double> vector_sum(const std::vector<double> &A,
 
 void scale_vector(double scalar, std::vector<double> &vec);
 
-int get_prev_phi(int phi);
+int get_prev_phi(double phi);
 
-int get_nearest_phi(int phi);
+int get_nearest_phi(double phi);
 
-int get_next_phi(int phi);
+int get_next_phi(double phi);
 
-int get_prev_theta(int theta, int phi);
+int get_prev_theta(double theta, double phi);
 
-int get_nearest_theta(int theta, int phi);
+int get_nearest_theta(double theta, double phi);
 
-int get_next_theta(int theta, int phi);
+int get_next_theta(double theta, double phi);
 
 #endif //CPP_INTERPOLATION_H
